# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Payment processing backend API with health monitoring.",
        "title": "Cashier Backend API",
        "termsOfService": "http://example.com/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8888",
    "basePath": "/",
    "paths": {
        "/api/v1/admin/get_membership_statistic": {
            "post": {
                "description": "Retrieves daily membership statistics.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Get Membership Statistics (Admin)",
                "parameters": [
                    {
                        "description": "Statistic request parameters",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/statistics.MembershipStatisticRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.RespMembershipStatistic"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/list_user_membership_item": {
            "post": {
                "description": "Retrieves a paginated and filterable list of all membership transactions.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "List Membership Transactions (Admin)",
                "parameters": [
                    {
                        "description": "List transaction request with filters, pagination, and sorting",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.ListTransactionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.RespListMembershipTransactions"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/send_free_gift": {
            "post": {
                "description": "Grants a free membership item to a user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Send Free Gift (Admin)",
                "parameters": [
                    {
                        "description": "Send free gift request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/transaction.SendFreeGiftRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.RespOK"
                        }
                    }
                }
            }
        },
        "/api/v1/user/transaction/list": {
            "get": {
                "description": "Retrieves a paginated list of transactions for a specific user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "List User Transactions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Pagination offset",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "Pagination limit",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "purchase_at",
                        "description": "Sort by field",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "default": "desc",
                        "description": "Sort order (asc/desc)",
                        "name": "sort_order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.RespUserListTransactions"
                        }
                    }
                }
            }
        },
        "/api/v1/user/transaction/verify": {
            "post": {
                "description": "Verifies a payment transaction with the provider and grants entitlement.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Verify Transaction",
                "parameters": [
                    {
                        "description": "Transaction verification request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/transaction.TransactionVerifyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful verification",
                        "schema": {
                            "$ref": "#/definitions/handlers.RespOK"
                        }
                    }
                }
            }
        },
        "/api/v1/webhook/apple": {
            "post": {
                "description": "Handles App Store Server Notifications V2. The request body should be a Signed JWS payload.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Webhook"
                ],
                "summary": "Apple Webhook",
                "parameters": [
                    {
                        "description": "App Store Server Notification V2 JWS payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.RespOK"
                        }
                    }
                }
            }
        },
        "/healthz": {
            "get": {
                "description": "Returns service status",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.ListMembershipTransactionsResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.TransactionItem"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "handlers.ListTransactionRequest": {
            "type": "object",
            "properties": {
                "filters": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.CommonFilter"
                    }
                },
                "from": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "sort_by": {
                    "type": "string"
                },
                "sort_order": {
                    "type": "string"
                }
            }
        },
        "handlers.RespListMembershipTransactions": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/response.APIResponseCode"
                },
                "data": {
                    "$ref": "#/definitions/handlers.ListMembershipTransactionsResponse"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "handlers.RespMembershipStatistic": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/response.APIResponseCode"
                },
                "data": {
                    "$ref": "#/definitions/statistics.MembershipStatisticResponse"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "handlers.RespOK": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/response.APIResponseCode"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "handlers.RespUserListTransactions": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/response.APIResponseCode"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.SwaggerTransaction"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "handlers.SwaggerTransaction": {
            "type": "object",
            "properties": {
                "auto_renew_expire_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "next_auto_renew_at": {
                    "type": "string"
                },
                "parent_transaction_id": {
                    "type": "string"
                },
                "payment_item_id": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "provider_id": {
                    "$ref": "#/definitions/types.PaymentProvider"
                },
                "purchase_at": {
                    "type": "string"
                },
                "refund_at": {
                    "type": "string"
                },
                "transaction_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "handlers.TransactionItem": {
            "type": "object",
            "properties": {
                "auto_renew_expire_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_first_purchase": {
                    "type": "boolean"
                },
                "membership_duration_minutes": {
                    "type": "integer"
                },
                "next_auto_renew_at": {
                    "type": "string"
                },
                "parent_transaction_id": {
                    "type": "string"
                },
                "payment_item_id": {
                    "type": "string"
                },
                "payment_item_type": {
                    "$ref": "#/definitions/types.PaymentItemType"
                },
                "price": {
                    "type": "integer"
                },
                "provider_id": {
                    "$ref": "#/definitions/types.PaymentProvider"
                },
                "provider_item_id": {
                    "type": "string"
                },
                "purchase_at": {
                    "type": "string"
                },
                "refund_at": {
                    "type": "string"
                },
                "transaction_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "response.APIResponseCode": {
            "type": "integer",
            "enum": [
                0,
                40000,
                50000
            ],
            "x-enum-varnames": [
                "APIResponseCodeOK",
                "APIResponseCodeBadRequest",
                "APIResponseCodeError"
            ]
        },
        "statistics.MembershipStatisticDataItem": {
            "type": "object",
            "properties": {
                "id": {
                    "$ref": "#/definitions/statistics.StatisticType"
                }
            }
        },
        "statistics.MembershipStatisticRequest": {
            "type": "object",
            "properties": {
                "data_items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/statistics.MembershipStatisticDataItem"
                    }
                },
                "filters": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.CommonFilter"
                    }
                }
            }
        },
        "statistics.MembershipStatisticResponse": {
            "type": "object",
            "properties": {
                "data_items": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "$ref": "#/definitions/statistics.MembershipStatisticResponseDataItem"
                        }
                    }
                }
            }
        },
        "statistics.MembershipStatisticResponseDataItem": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "value": {
                    "type": "integer"
                },
                "value2": {
                    "type": "integer"
                },
                "value3": {
                    "type": "integer"
                }
            }
        },
        "statistics.StatisticType": {
            "type": "string",
            "enum": [
                "daily_transaction_count",
                "daily_gmv",
                "total_gmv",
                "daily_membership_count",
                "daily_new_membership_count",
                "total_membership_count",
                "daily_accumulated_membership_count",
                "renewal_success_rate"
            ],
            "x-enum-varnames": [
                "StatisticTypeDailyTransactionCount",
                "StatisticTypeDailyGmv",
                "StatisticTypeTotalGmv",
                "StatisticTypeDailyMembershipCount",
                "StatisticTypeDailyNewMembershipCount",
                "StatisticTypeTotalMembershipCount",
                "StatisticTypeDailyAccumulatedMembershipCount",
                "StatisticTypeRenewalSuccessRate"
            ]
        },
        "transaction.SendFreeGiftRequest": {
            "type": "object",
            "properties": {
                "operator_id": {
                    "description": "操作员ID",
                    "type": "string"
                },
                "payment_item_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "transaction.TransactionVerifyRequest": {
            "type": "object",
            "properties": {
                "provider_id": {
                    "type": "string"
                },
                "transaction_id": {
                    "type": "string"
                }
            }
        },
        "types.CommonFilter": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "filters": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.CommonFilter"
                    }
                },
                "operator": {
                    "$ref": "#/definitions/types.CommonFilterOperator"
                },
                "values": {
                    "type": "array",
                    "items": {}
                }
            }
        },
        "types.CommonFilterOperator": {
            "type": "string",
            "enum": [
                "eq",
                "not_eq",
                "lt",
                "lte",
                "gt",
                "gte",
                "date_range",
                "range",
                "in"
            ],
            "x-enum-varnames": [
                "CommonFilterOperatorEq",
                "CommonFilterOperatorNotEq",
                "CommonFilterOperatorLt",
                "CommonFilterOperatorLte",
                "CommonFilterOperatorGt",
                "CommonFilterOperatorGte",
                "CommonFilterOperatorDateRange",
                "CommonFilterOperatorRange",
                "CommonFilterOperatorIn"
            ]
        },
        "types.PaymentItemType": {
            "type": "string",
            "enum": [
                "auto_renewable_subscription",
                "non_renewable_subscription"
            ],
            "x-enum-varnames": [
                "PaymentItemTypeAutoRenewableSubscription",
                "PaymentItemTypeNonRenewableSubscription"
            ]
        },
        "types.PaymentProvider": {
            "type": "string",
            "enum": [
                "apple",
                "google",
                "inner"
            ],
            "x-enum-varnames": [
                "PaymentProviderApple",
                "PaymentProviderGoogle",
                "PaymentProviderInner"
            ]
        }
    }
}